compose.__call__: fix missing-self error
Originally I skipped implementing this error because this error
requires going out of your way to use the method both weirdly
and incorrectly.

But ultimately, people might be relying on standard error
behavior. And if you are doing something weird and hit an
error, it is nicer if the error is as close to obvious as
possible, as close to standard as possible.