Added sacompose
Most code should just use `compose` or `acompose`, because most code
has no good reason to not statically "know" if it is `async` or not.

But interactive use and operator overloads can benefit from dynamic
selection of synchronous or asynchronous behavior at runtime, and
there might be other situations that I am not thinking of as well.