Streamline/trim description of `sacompose`
Removed the proscriptive parts entirely.
Possibly to the point of losing some of
the descriptiveness (after all, "this is
good for [use-case]" can be partially or
entirely descriptive).

Based on my experience coming back to this
description since I wrote it,

1. it was not obvious why both `acompose` and
   `sacompose` were being mentioned at the
   beginning, since the difference didn't get
   explained until a couple paragraphs/examples
   later, and
2. the introduction of the difference between
   them (the corner case where it matters) was
   perceptibly effortful/errorprone to parse.

In the effort to trim that down, I've decided to
err on the side of conciseness and trusting the
reader to pick the one they need once they know
the differences (and the other kind of user will
just use the first "good enough" one, and I've
implicitly suggested `acompose` over `sacompose`
as the default by mentioning it first and more).